requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chex==0.1.7
cmake==3.26.1
contourpy==1.0.7
cycler==0.11.0
dm-tree==0.1.8
etils==1.1.1
filelock==3.10.7
flatbuffers==23.3.3
flax==0.6.8
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
jax==0.4.8
jaxlib==0.4.7
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.0.4
mpmath==1.3.0
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.0
discord~=2.2.2
python-dotenv~=1.0.0
numpy~=1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.7
packaging==23.0
Pillow==9.5.0
pip-review==1.3.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.9
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorstore==0.1.35
termcolor==2.2.0
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt
sacremoses
transformers~=4.27.4